services/signature/pyproject.toml
[tool.poetry]
name = "signature-service"
version = "0.1.0"
description = "Document signature and authentication service"
authors = []

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
azure-servicebus = "^7.11.4"
pydantic = "^2.12.3"
pydantic-settings = "^2.6.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
psycopg = {extras = ["binary"], version = "^3.1.0"}
asyncpg = "^0.30.0"
httpx = "^0.26.0,<0.29.0"
cryptography = "^41.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = {extras = ["hiredis"], version = "^5.0.1"}
tenacity = "^8.2.3"
carpeta-common = {path = "../common", develop = true}

[tool.poetry.group.dev.dependencies]
pytest-cov = "^7.0.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-httpx = "^0.28.0"
fakeredis = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

